Docker/yahoo_llm_project/viz_service/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
werkzeug==2.3.7
